refactor(Vehicle): 将车辆的电量值类型修改为double类

+++ openTCS-API-Base/src/main/java/org/opentcs/components/kernel/services/InternalVehicleService.java
@@ -32,7 +32,7 @@ public interface InternalVehicleService
    * @param energyLevel The vehicle's new energy level.
    * @throws ObjectUnknownException If the referenced vehicle does not exist.
    */
-  void updateVehicleEnergyLevel(TCSObjectReference<Vehicle> ref, int energyLevel)
+  void updateVehicleEnergyLevel(TCSObjectReference<Vehicle> ref, double energyLevel)
       throws ObjectUnknownException;
 
   /**
